Exception Handling - delete

+++ src/main/java/com/racalixta/course/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.racalixta.course.entities.User;
 import com.racalixta.course.repositories.UserRepository;
+import com.racalixta.course.services.exceptions.DatabaseException;
 import com.racalixta.course.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -29,9 +31,16 @@ public class UserService {
 		return repository.save(obj);
 	}
 	
-	public void delete(Long id) {
-		repository.deleteById(id);
-	}
+    public void delete(Long id) {
+        try {
+        	if (!repository.existsById(id)) {
+        		throw new ResourceNotFoundException(id);
+        	}
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+        	throw new DatabaseException(e.getMessage());
+        }
+    }
 	
 	public User update(Long id, User obj) {
 		User entity = repository.getReferenceById(id);
